add ckpt_source flag to loraRequest, simplify tests
