+++ tensorrt_llm/executor/request.py
@@ -25,10 +25,15 @@ class LoRARequest:
     lora_name: str
     lora_int_id: int
     lora_path: str = ""
+    lora_ckpt_source: str = "hf"
 
     def __post_init__(self):
         if self.lora_path is not None and not os.path.exists(self.lora_path):
             raise ValueError(f"lora_path ({self.lora_path}) does not exist.")
+        if self.lora_ckpt_source not in ["hf", "nemo"]:
+            raise ValueError(
+                f"lora_ckpt_source must be 'hf' or 'nemo', got '{self.lora_ckpt_source}'"
+            )
 
     @property
     def adapter_id(self):
@@ -41,6 +46,10 @@ class LoRARequest:
     @property
     def path(self):
         return self.lora_path
+
+    @property
+    def ckpt_source(self):
+        return self.lora_ckpt_source
 
 
 @dataclass(slots=True)
