+++ tests/unittest/llmapi/test_llm_pytorch.py
@@ -29,7 +29,102 @@ from peft import LoraConfig as PeftLoraConfig
 from peft import get_peft_model
 from transformers import AutoModelForCausalLM
 
+import json
+import tarfile
+from pathlib import Path
+
 # isort: on
+
+# NeMo LoRA test data
+LORA_RANK_CONFIGS = [
+    # (lora_rank, max_lora_rank, description)
+    (8, 8, "rank_8"),
+    (16, 16, "rank_16"),
+    (4, 8, "rank_4_max_8"),
+]
+
+
+def create_mock_nemo_lora_checkpoint(
+    lora_dir: Path,
+    hidden_size: int = 4096,
+    num_layers: int = 32,
+    lora_rank: int = 8,
+    tp_size: int = 1,
+) -> Path:
+    """Create a minimal NeMo LoRA checkpoint for testing.
+
+    This creates a .nemo tarfile with the expected structure:
+    - model_weights.ckpt containing attn_qkv adapter weights
+    - model_config.yaml with basic configuration
+
+    Args:
+        lora_dir: Directory to create the checkpoint in
+        hidden_size: Model hidden size
+        num_layers: Number of transformer layers
+        lora_rank: LoRA rank
+        tp_size: Tensor parallelism size
+
+    Returns:
+        Path to the created .nemo file
+    """
+    # Create temporary directory for checkpoint contents
+    temp_dir = lora_dir / "temp_nemo"
+    temp_dir.mkdir(exist_ok=True)
+
+    # Create LoRA weights dict
+    weights_dict = {}
+
+    for layer_idx in range(num_layers):
+        # NeMo uses this key format for QKV adapters
+        key_prefix = f"model.layers.{layer_idx}.self_attention.adapter_layer.lora_kqv_adapter"
+
+        # Create linear_in weights [lora_rank, hidden_size] with small random values
+        linear_in_key = f"{key_prefix}.linear_in.weight"
+        weights_dict[linear_in_key] = torch.randn(
+            lora_rank, hidden_size, dtype=torch.float16) * 0.01
+
+        # Create linear_out weights [3 * hidden_size, lora_rank] for QKV combined
+        linear_out_key = f"{key_prefix}.linear_out.weight"
+        weights_dict[linear_out_key] = torch.randn(
+            3 * hidden_size, lora_rank, dtype=torch.float16) * 0.01
+
+    # Save checkpoint
+    ckpt_path = temp_dir / "model_weights.ckpt"
+    torch.save(weights_dict, ckpt_path)
+
+    # Create minimal config
+    config = {
+        "precision": "fp16",
+        "trainer": {
+            "num_nodes": 1,
+            "devices": tp_size,
+        },
+        "model": {
+            "hidden_size": hidden_size,
+            "num_layers": num_layers,
+        },
+        "lora": {
+            "rank": lora_rank,
+            "target_modules": ["attn_qkv"],
+        }
+    }
+
+    config_path = temp_dir / "model_config.yaml"
+    # Using JSON for simplicity since YAML parsing isn't critical for the test
+    with open(config_path, 'w') as f:
+        json.dump(config, f)
+
+    # Create .nemo tarfile
+    nemo_path = lora_dir / "test_lora.nemo"
+    with tarfile.open(nemo_path, 'w') as tar:
+        tar.add(ckpt_path, arcname="model_weights.ckpt")
+        tar.add(config_path, arcname="model_config.yaml")
+
+    # Cleanup temp dir
+    import shutil
+    shutil.rmtree(temp_dir)
+
+    return nemo_path
 
 
 @force_ampere
@@ -427,3 +522,163 @@ def test_bielik_11b_v2_2_instruct_multi_lora() -> None:
                                lora_request=lora_requests)
 
         assert len(outputs) == 2
+
+
+# NeMo LoRA tests
+@pytest.mark.parametrize("lora_rank,max_lora_rank,description",
+                         LORA_RANK_CONFIGS)
+def test_load_torch_nemo_lora_function(lora_rank, max_lora_rank, description):
+    """Test load_torch_nemo_lora function with different LoRA rank configurations."""
+    from tensorrt_llm.lora_manager import load_torch_nemo_lora
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        temp_path = Path(temp_dir)
+
+        # Create a mock NeMo checkpoint
+        nemo_path = create_mock_nemo_lora_checkpoint(
+            temp_path,
+            hidden_size=2048,
+            num_layers=16,
+            lora_rank=lora_rank,
+        )
+
+        # Test load_torch_nemo_lora
+        lora_config = LoraConfig(
+            lora_dir=[str(nemo_path)],
+            lora_ckpt_source="nemo",
+            max_lora_rank=max_lora_rank,
+        )
+
+        # This should not raise an error
+        load_torch_nemo_lora(lora_config)
+
+        # Verify configuration was set correctly
+        assert lora_config.lora_target_modules == [
+            "attn_qkv"
+        ], f"Expected attn_qkv modules for {description}"
+        assert lora_config.trtllm_modules_to_hf_modules == {
+            "attn_qkv": "attn_qkv"
+        }, f"Expected correct module mapping for {description}"
+
+
+def test_nemo_lora_unsupported_modules_validation():
+    """Test validation of unsupported modules in NeMo LoRA."""
+    from tensorrt_llm.lora_manager import load_torch_nemo_lora
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        temp_path = Path(temp_dir)
+
+        # Create a mock NeMo checkpoint
+        nemo_path = create_mock_nemo_lora_checkpoint(
+            temp_path,
+            hidden_size=2048,
+            num_layers=16,
+            lora_rank=8,
+        )
+
+        # Test validation: should fail with unsupported modules
+        invalid_config = LoraConfig(
+            lora_dir=[str(nemo_path)],
+            lora_ckpt_source="nemo",
+            lora_target_modules=["attn_qkv",
+                                 "mlp_h_to_4h"],  # mlp_h_to_4h not supported
+            max_lora_rank=8,
+        )
+
+        with pytest.raises(ValueError, match="NeMo LoRA only supports"):
+            load_torch_nemo_lora(invalid_config)
+
+
+@force_ampere
+def test_tinyllama_nemo_lora():
+    """Test end-to-end generation with NeMo LoRA checkpoint."""
+    with tempfile.TemporaryDirectory() as temp_dir:
+        temp_path = Path(temp_dir)
+
+        # Create a mock NeMo checkpoint for TinyLlama
+        # TinyLlama has hidden_size=2048, num_layers=22
+        nemo_path = create_mock_nemo_lora_checkpoint(
+            temp_path,
+            hidden_size=2048,
+            num_layers=22,
+            lora_rank=8,
+        )
+
+        # Create LoRA config for NeMo checkpoint
+        lora_config = LoraConfig(
+            lora_dir=[str(nemo_path)],
+            lora_ckpt_source="nemo",
+            max_lora_rank=8,
+        )
+
+        # Verify LoRA config is set up correctly
+        assert lora_config.lora_ckpt_source == "nemo"
+        assert len(lora_config.lora_dir) == 1
+        print(f"✓ Created NeMo LoRA config: {nemo_path}")
+
+        # Use TinyLlama for fast testing
+        model_path = get_model_path("llama-models-v2/TinyLlama-1.1B-Chat-v1.0")
+
+        # Create LLM with NeMo LoRA
+        llm = LLM(
+            model=model_path,
+            lora_config=lora_config,
+            kv_cache_config=global_kvcache_config,
+        )
+
+        try:
+            # Test prompts
+            test_prompts = [
+                "Hello, how are you?",
+                "What is the capital of France?",
+            ]
+
+            # Create LoRA request for the NeMo checkpoint
+            lora_req = LoRARequest("nemo-task",
+                                   0,
+                                   str(nemo_path),
+                                   lora_ckpt_source="nemo")
+
+            # Verify LoRA request is configured correctly
+            assert lora_req.ckpt_source == "nemo"
+            assert lora_req.path == str(nemo_path)
+
+            # Test with and without LoRA
+            sampling_params = SamplingParams(max_tokens=20, temperature=0.0)
+
+            # Generate with LoRA
+            outputs_with_lora = llm.generate(test_prompts,
+                                             sampling_params,
+                                             lora_request=[lora_req, lora_req])
+
+            # Generate without LoRA
+            outputs_without_lora = llm.generate(test_prompts,
+                                                sampling_params,
+                                                lora_request=[None, None])
+
+            # Basic validation - outputs should be generated without errors
+            assert len(outputs_with_lora) == 2
+            assert len(outputs_without_lora) == 2
+
+            # Verify that generation completed successfully (may have minimal output with mock weights)
+            for i in range(2):
+                # Check that we got valid completion outputs
+                assert outputs_with_lora[i].outputs[0] is not None
+                assert outputs_without_lora[i].outputs[0] is not None
+                # Check that token_ids are present (even if just EOS token)
+                assert len(outputs_with_lora[i].outputs[0].token_ids) > 0
+                assert len(outputs_without_lora[i].outputs[0].token_ids) > 0
+
+            print(f"✓ NeMo LoRA generation completed successfully")
+            print(
+                f"✓ LoRA output tokens: {[len(out.outputs[0].token_ids) for out in outputs_with_lora]}"
+            )
+            print(
+                f"✓ Base output tokens: {[len(out.outputs[0].token_ids) for out in outputs_without_lora]}"
+            )
+
+            # Test passes if generation completes without errors
+            # Note: With mock LoRA weights, outputs may be minimal but that's expected
+
+        finally:
+            llm.shutdown()
